requirements.txt
azure-ai-inference==1.0.0b8
discord.py==2.4.0
google-generativeai==0.8.3
loguru==0.7.2
pillow==11.1.0
python-dotenv==1.0.1
pytubefix==8.12.1
pynacl==1.5.0
requests==2.32.3
youtube-search==2.1.2
google-api-python-client==2.154.0